+++ unreal_asset/src/properties/movies/movie_scene_segment_property.rs
@@ -3,8 +3,9 @@ use byteorder::LittleEndian;
 use crate::{
     error::Error,
     impl_property_data_trait, optional_guid, optional_guid_write,
-    properties::{core_uobject::FFrameNumberRange, Property, PropertyTrait},
+    properties::{Property, PropertyTrait},
     reader::{asset_reader::AssetReader, asset_writer::AssetWriter},
+    types::movie::FFrameNumberRange,
     types::{FName, Guid},
 };
